application completed

+++ controllers/contactController.js
@@ -3,7 +3,7 @@ const asyncHandler = require('express-async-handler');
 const Contact = require('../models/contactModel');
 
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json(contacts);
 });
 
@@ -22,7 +22,10 @@ const deleteContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error('Contact not found');
   }
-
+  if (contact.user_id.toString() != req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to delete other user contacs");
+  }
   await Contact.deleteOne({ _id: req.params.id });
   res.status(200).json(contact);
 });
@@ -38,6 +41,7 @@ const createContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
   res.status(200).json(contact);
 });
@@ -47,6 +51,11 @@ const updateContact = asyncHandler(async (req, res) => {
   if (!contact) {
     res.status(404);
     throw new Error('Contact not found');
+  }
+
+  if (contact.user_id.toString() != req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to update other user contacs");
   }
 
   const updatedContact = await Contact.findByIdAndUpdate(
